pyproject.toml
[tool.poetry]
name = "webforumapp"
version = "0.1.0"
description = ""
authors = ["Mateusz Samek <samekmat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
